// src/integrations/supabase/types.ts
export type Database = {
  public: {
    Tables: {
// ...
      seaports: {
        Row: {
          id: string;
          name: string;
          address: string | null;
          classification: number | null;
          province: string | null;
          district: string | null;
          ward: string | null;
          status: string | null;
          latitude: number | null;
          longitude: number | null;
          image_url: string | null;
          created_at: string | null;
          updated_at: string | null;
          number: number;
        };
        Insert: {
          id?: string;
          name: string;
          address?: string | null;
          classification?: number | null;
          province?: string | null;
          district?: string | null;
          ward?: string | null;
          status?: string | null;
          latitude?: number | null;
          longitude?: number | null;
          image_url?: string | null;
          created_at?: string | null;
          updated_at?: string | null;
          number?: number;
        };
        Update: {
          id?: string;
          name?: string;
          address?: string | null;
          classification?: number | null;
          province?: string | null;
          district?: string | null;
          ward?: string | null;
          status?: string | null;
          latitude?: number | null;
          longitude?: number | null;
          image_url?: string | null;
          created_at?: string | null;
          updated_at?: string | null;
          number?: number;
        };
      };
      users: {
        Row: {
          id: string;
          auth_id: string | null;
          email: string;
          name: string;
          is_approved: boolean | null;
          is_email_verified: boolean | null;
          is_phone_verified: boolean | null;
          phone: string | null;
          created_at: string | null;
          updated_at: string | null;
          grant_id: string | null;
          role: string | null;
          account_type: string | null;
          account_package: string | null;
          default_vessel: string | null;
        };
        Insert: {
          id?: string;
          auth_id?: string | null;
          email: string;
          name: string;
          is_approved?: boolean | null;
          is_email_verified?: boolean | null;
          is_phone_verified?: boolean | null;
          phone?: string | null;
          created_at?: string | null;
          updated_at?: string | null;
          grant_id?: string | null;
          role?: string | null;
          account_type?: string | null;
          account_package?: string | null;
          default_vessel?: string | null;
        };
        Update: {
          id?: string;
          auth_id?: string | null;
          email?: string;
          name?: string;
          is_approved?: boolean | null;
          is_email_verified?: boolean | null;
          is_phone_verified?: boolean | null;
          phone?: string | null;
          created_at?: string | null;
          updated_at?: string | null;
          grant_id?: string | null;
          role?: string | null;
          account_type?: string | null;
          account_package?: string | null;
          default_vessel?: string | null;
        };
      };
// ...
    };
    Views: Record<string, never>;
    Functions: Record<string, never>;
    Enums: Record<string, never>;
  };
};